requirements.txt
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
fonttools==4.61.1
fpdf2==2.8.5
google-auth==2.47.0
google-genai==1.57.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
markdown2==2.5.4
openai==2.15.0
pillow==12.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydyf==0.12.1
pyphen==0.17.2
python-dotenv==1.2.1
python-pptx==1.0.2
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
weasyprint==67.0
webencodings==0.5.1
websockets==15.0.1
xlsxwriter==3.2.9
zopfli==0.4.0
